requirements.txt
anthropic>=0.42.0
openai>=1.60.0
google-genai>=1.0.0
pyyaml>=6.0
rich>=13.0
python-dotenv>=1.0
httpx>=0.27.0
PyJWT>=2.8.0
moviepy>=2.0.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0
edge-tts>=7.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
